✨(frontend) add a TimedText pane to let users manage tracks in Dashboard
Add a DashboardTimedTextPane next to the DashboardVideoPane in the
Dashboard to let users create new timedtexttracks, see the state
of existing ones, delete them or replace the related files.

We split this Pane in a series of simpler components responsible for
the parts of the UI. One of our structuring decisions was to split
the tracks by mode in the interface to make it easy for users to
understand and for us to enforce our video-mode-language unicity
rule for timedtexttracks.

We also added the side effects we needed as separate async functions,
as adding a proper side effect management library was out of the scope
of this feature. These can easily be refactored to use `saga` or
`thunks` when we decide to structure this part of the app.